Make outgoings service to divide view and logic
